[test] document destroy tests
git-svn-id: svn://svn.zenadmin.org/zena/trunk@587 aaa19a3b-e121-0410-9d45-962fd44e537e

+++ lib/multiversion.rb
@@ -204,9 +204,7 @@ module Zena
             end
           when :destroy_version
             if versions.count == 1
-              debugger
-              version.destroy 
-              self.destroy
+              version.destroy && self.destroy
             else
               version.destroy
             end
